Create a Boilerplate R Package.

Creates a boilerplate R package by generating the following:

DESCRIPTION file (see jeksterslabRpkg::pkg_description()),

NAMESPACE file (see jeksterslabRpkg::pkg_namespace()),

LICENSE and LICENSE.md (see jeksterslabRpkg::pkg_license()),

a simple R script and accompanying Rd documentation file (see jeksterslabRpkg::pkg_r() and jeksterslabRpkg::pkg_rd()),

sample data, test, and vignette (see jeksterslabRpkg::pkg_data(), (see jeksterslabRpkg::pkg_test(), and jeksterslabRpkg::pkg_vignette()),

optional pkgdown and travis YAML files (see jeksterslabRpkg::pkg_pkgdown() and jeksterslabRpkg::pkg_travis()),

optional README.Rmd file (see jeksterslabRpkg::pkg_readme()), and

other files like .Rbuildignore, .gitignore, and Rproj.

pkg_create(
  pkg_dir = getwd(),
  pkg_name = "boilerplatePackage",
  input_file = NULL,
  pkgdown = FALSE,
  travis = FALSE,
  appveyor = FALSE,
  readme = FALSE,
  add_description = NULL,
  add_namespace = NULL,
  add_rbuildignore = NULL,
  add_gitignore = NULL,
  add_travis = NULL,
  add_appveyor = NULL,
  git = FALSE,
  github = FALSE
)

Arguments

pkg_dir
Character string. Directory where the package is initialized.

pkg_name
Character string. Package name.

input_file
Character string. csv file containing DESCRIPTION fields and entries.

pkgdown
Logical. Create pkgdown YAML file.

travis
Logical. Create travis YAML file.

appveyor
Logical. Create appveyor YAML file.

readme
Logical. Create README.Rmd file.

add_description
Character string. Additional entries to the DESCRIPTION file not included in input_file.

add_namespace
Character string. Entries to the NAMESPACE in addition to the boilerplate example.

add_rbuildignore
Character string. Entries to the .Rbuildignore in addition to the boilerplate example.

add_gitignore
Character string. Entries to the .gitignore in addition to the boilerplate example.

add_travis
Character string. Entries to the .travis.yml in addition to the boilerplate example.

add_appveyor
Character string. Entries to the appveyor.yml in addition to the boilerplate example.

git
Logical. Set up a git repository.

github
Logical. Set up and push to a github repository.

Details

Optionally, the function can also set up a Git repository and push the created repo to Github. This requires that git and hub are installed and configured in the system.

If you are going to document, check, and build your package using devtools, you may delete the boillerplate NAMESPACE and MAN files.

Examples

if (FALSE) {
pkg_create(
  pkg_dir = getwd(),
  pkg_name = "boilerplatePackage",
  input_file = "DESCRIPTION.csv",
  git = TRUE,
  github = TRUE
)
}