Alias for devtools::check()

Alias for devtools::check()

rchk(
  pkg = ".",
  document = NA,
  build_args = NULL,
  ...,
  manual = FALSE,
  cran = TRUE,
  remote = FALSE,
  incoming = remote,
  force_suggests = FALSE,
  run_dont_test = FALSE,
  args = "--timings",
  env_vars = c(NOT_CRAN = "true"),
  quiet = FALSE,
  check_dir = tempdir(),
  cleanup = TRUE,
  vignettes = TRUE,
  error_on = c("never", "error", "warning", "note")
)

Arguments

pkg
The package to use, can be a file path to the package or a package object. See as.package() for more information.

document
If NA and the package uses roxygen2, will rerun document() prior to checking. Use TRUE and FALSE to override this default.

build_args
Additional arguments passed to R CMD build

...
Additional arguments passed on to pkgbuild::build().

manual
If FALSE, don't build and check manual (--no-manual).

cran
if TRUE (the default), check using the same settings as CRAN uses.

remote
Sets _R_CHECK_CRAN_INCOMING_REMOTE_ env var. If TRUE, performs a number of CRAN incoming checks that require remote access.

incoming
Sets _R_CHECK_CRAN_INCOMING_ env var. If TRUE, performs a number of CRAN incoming checks.

force_suggests
Sets _R_CHECK_FORCE_SUGGESTS_. If FALSE (the default), check will proceed even if all suggested packages aren't found.

run_dont_test
Sets --run-donttest so that tests surrounded in \dontest{} are also tested. This is important for CRAN submission.

args
Character vector of arguments to pass to R CMD check. (Note that instead of the --output option you should use the check_dir argument, because --output cannot deal with spaces and other special characters on Windows.

env_vars
Environment variables set during R CMD check

quiet
if TRUE suppresses output from this function.

check_dir
the directory in which the package is checked compatibility. args = "--output=/foo/bar" can be used to change the check directory.

cleanup
Deprecated.

vignettes
If FALSE, do not build or check vignettes, equivalent to using args = '--ignore-vignettes' and build_args = '--no-build-vignettes'.

error_on
Whether to throw an error on R CMD check failures. Note that the check is always completed (unless a timeout happens), and the error is only thrown after completion. If "never", then no errors are thrown. If "error", then only ERROR failures generate errors. If "warning", then WARNING failures generate errors as well. If "note", then any check failure generated an error.

See also