Alias for styler::style_dir()

Alias for styler::style_dir()

rsty(
  path = ".",
  ...,
  style = styler::tidyverse_style,
  transformers = style(...),
  filetype = c("R", "Rprofile"),
  recursive = TRUE,
  exclude_files = NULL,
  exclude_dirs = c("packrat", "renv"),
  include_roxygen_examples = TRUE
)

Arguments

path
Path to a directory with files to transform.

...
Arguments passed on to the style function.

style
A function that creates a style guide to use, by default tidyverse_style() (without the parentheses). Not used further except to construct the argument transformers. See style_guides() for details.

transformers
A set of transformer functions. This argument is most conveniently constructed via the style argument and .... See 'Examples'.

filetype
Vector of file extensions indicating which file types should be styled. Case is ignored, and the . is optional, e.g. c(".R", ".Rmd"), or c("r", "rmd"). Supported values (after standardization) are: "r", "rprofile", "rmd", "rnw".

recursive
A logical value indicating whether or not files in subdirectories of path should be styled as well.

exclude_files
Character vector with paths to files that should be excluded from styling.

exclude_dirs
Character vector with directories to exclude.

include_roxygen_examples
Whether or not to style code in roxygen examples.

See also